requirements.txt
numpy
pandas
scikit-learn
matplotlib
seaborn
requests
flask
kafka-python
streamlit
kaggle